% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region-reorg-fcns.R
\name{abv.rcols}
\alias{abv.rcols}
\title{These functions just manipulate dataframes that contain multiple
regions/regiontypes (i.e., CZs and CBSAs). Some may have general broad use, others
may just help me shift quickly between conventions that I've developed for this
sort of data.
abv.rcols}
\usage{
abv.rcols(x)
}
\arguments{
\item{x}{df with columns with name to possibly abbreviate}
}
\description{
Shortens identifier column names.
\itemize{
\item region.name -> rn
}
}
\details{
\itemize{
\item region.id -> rid
\item region.type -> rt
}
}
