% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialize-fcns.R
\name{attach.geos}
\alias{attach.geos}
\title{attach.geos}
\usage{
attach.geos(
  x,
  geoid.col = "geoid",
  query.fcn = tigris::tracts,
  crs = 4326,
  keep.census.cols = F,
  ...
)
}
\arguments{
\item{x}{df with at least 1 column containing state+county geoid hierarchy (so
that first 5 characters are state and county fips codes)}

\item{geoid.col}{column representing geoids}

\item{query.fcn}{function to query geos from census api. Default gets tracts}

\item{keep.census.cols}{Whether to keep all columns from query, like aland,
awater, etc. If false, just gets geometry.}

\item{...}{passed onto the \code{query.fcn}}
}
\value{
\code{x}, as \code{sf} object with geometries attached
}
\description{
attach.geos
}
