% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{county.subset}
\alias{county.subset}
\title{functions that interface with the \code{tigris} library to query layers.}
\usage{
county.subset(x, cos = NULL, ...)
}
\arguments{
\item{x}{\code{sf} or \code{bbox} object spanning area you want to filter polys to. Will use
get intersection if sf object or crop if bbox.}

\item{cos}{counties sf object. If none supplied, they are retrieved using
\code{tigris}/census api}

\item{...}{passed onto \code{tigris::counties}}
}
\description{
The flow is: given a spatial layer, can query counties with county_subset, a thin
wrapper. Then other thin wrappers will call
county.subset
}
\details{
Given an \code{sf} object or bbox, get all counties that fall within polygon or bbox.
}
