% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosubset-fcns.R
\name{geosubset}
\alias{geosubset}
\title{geosubset}
\usage{
geosubset(x, subset.cols = c("geoid"), ...)
}
\arguments{
\item{x}{df with at least 1 column containing state+county geoid hierarchy}

\item{subset.cols}{columns to subset to. \code{geoid} by default; \code{c('origin','dest')}
may be useful for mobility data}

\item{cz_name, cz, countyfp, cbsa_id, plc_id}{Region/county/place to subset to}
}
\value{
df \code{x}, subsetted to the specified region.
}
\description{
Subsets a dataframe with one column representing tracts, counties, census block
groups, or another areal unit with STATE+COUNTY fips codes comprising the first
5-characters of their geoid (as detailed here
https://www.census.gov/programs-surveys/geography/guidance/geo-identifiers.html).
}
\details{
Can take cz or cbsa or place to subset to, and use one of the xwalks in this
package to determine correspondence. In this case, attention should be payed to
the year of the areal unit.
}
