% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region-reorg-fcns.R
\name{get.region.identifiers}
\alias{get.region.identifiers}
\title{get.region.identifiers}
\usage{
get.region.identifiers(cz = NULL, cbsa = NULL, abvcols = T, remove.slashes = T)
}
\arguments{
\item{cz, cbsa}{one of a cz or cbsa identifier code (either 5-digit # or 5-char
numeric)}

\item{abvcols}{whether to abbreviate region id columns}

\item{remove.slashes}{Sometimes there are slashes ('/') in cbsa names. Replaces
with dashes if this is true.}
}
\value{
a 1-row tibble that organizes the region id/name/type
}
\description{
Gets region.id/region.name and attaches to region.type, to get bundled region id
information in format expected by other functions.
}
