% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{parks.wrapper}
\alias{parks.wrapper}
\title{parks.wrapper}
\usage{
parks.wrapper(x = NULL, .statefps = NULL, ...)
}
\description{
Gets parks in a state, based on census bureau data. Can query from sf object x or
state fp codes.
}
