% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{places.wrapper}
\alias{places.wrapper}
\title{places.wrapper}
\usage{
places.wrapper(x = NULL, .statefps = NULL, ...)
}
\arguments{
\item{x}{\code{sf} or \code{bbox} object spanning area you want to filter polys to. Will use
get intersection if sf object or crop if bbox.}

\item{.statefps}{statefps to get areas for}

\item{...}{passed onto \code{tigris::places}}
}
\description{
Gets parks in a state, based on census bureau data. Can query from sf object x or
state fp codes.
}
