% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpops}
\alias{rpops}
\title{Region populations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1667 rows and 3 columns.
}
\usage{
rpops
}
\description{
CZs and CBSAs (long by region type) with 2019 population attached.
}
\keyword{datasets}
