% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rx}
\alias{rx}
\title{CBSA/CZ/county region crosswalk}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3142 rows and 6 columns.
}
\usage{
rx
}
\description{
Crosswalk or areal correspondence from counties to CZs and CBSAs
1990 CZs and 2019 counties/cbsas
}
\keyword{datasets}
