% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{tracts.from.sf}
\alias{tracts.from.sf}
\title{tracts.from.sf}
\usage{
tracts.from.sf(x = NULL, countyfps = NULL, query.fcn = tigris::tracts, ...)
}
\description{
From an \code{sf} object, gets overlapping tracts or block groups.
}
