% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{trim.tigris.waters}
\alias{trim.tigris.waters}
\title{functions that interface with the \code{tigris} library to query layers.}
\usage{
trim.tigris.waters(
  wtr,
  mtfccs = c("H2051", "H2053", "H3010", "H2040"),
  size.floor = 1e+05
)
}
\arguments{
\item{wtr}{Water simple features, as from tigris::water, with lowercase
column names}

\item{mtfccs}{census mtfcc codes to retain}

\item{size.floor}{A catch-all size filter to retain larger features. 100,000
square meters by default.}
}
\description{
The flow is: given a spatial layer, can query counties with county_subset, a thin
wrapper. Then other thin wrappers will call
trim.tigris.waters
}
\details{
Tigris returns many very small water areas--- for most purposes, retaining
only a subset of these water areas is preferable. This function applies
filters based on census mtfcc code, feature name, and size to subset
features.

For list of MTFF codes:
https://www2.census.gov/geo/pdfs/reference/mtfccs2022.pdf
}
