% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{county.subset}
\alias{county.subset}
\title{county.subset}
\usage{
county.subset(x, cos = NULL, ...)
}
\arguments{
\item{x}{\code{sf} or \code{bbox} object spanning area you want to filter polys to. Will use
get intersection if sf object or crop if bbox.}

\item{cos}{counties sf object. If none supplied, they are retrieved using
\code{tigris}/census api}

\item{...}{passed onto \code{tigris::counties}}
}
\description{
Given an \code{sf} object or bbox, get all counties that fall within polygon or bbox.
}
