% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-fcns.R
\name{flexible.spatial.filter}
\alias{flexible.spatial.filter}
\title{flexible.spatial.filter}
\usage{
flexible.spatial.filter(x, polys)
}
\arguments{
\item{x}{\code{sf} or \code{bbox} object spanning area you want to filter polys to. Will use
get intersection if sf object or crop if bbox.}

\item{polys}{polygons to get over area \code{x}.}

\item{subset.approach}{spatial filter approach; one of "intersects" or "crop".
Intersection or cropping to bbox.}
}
\description{
Gets all polygons \code{polys} that overlap with sf object \code{x}.
}
\details{
May use different subset approaches (intersects or crop), depending on whether you
want to match a polygon's area or its bounding box.

If \code{x} isn't already points, it is transformed to them using
\code{st_points_on_surface}.

Currently this doesn't work with non-coterminous polys. In this case, use
\code{xwalks::get.spatial.overlap}, and filter by degree of overlap.
}
